Manually Draw the child views
A view with elements that have a visibility to gone on the initial render won't be displayed after you've set its visibility to visible. view.isShown() will return true, but it will not be there or it will be there but not really re-layout. This hack somehow draws the child views manually and fixes the issue. This solves both DRT-44 and DRT-47
For more info: https://github.com/facebook/react-native/issues/17968