Unregister event listeners in onDetachedFromWindow too
I noticed that if you are on the video screen then tapped on the back button or the toolbar up button, then onHostPasued doesn't get called, and the user is redirected to the home screen.
This results in an unwanted behavior, lets say you had a video playing, then turn off wifi, the click then go to the home screen, turn wifi on again, then you will start listening to the video playback, though you are on the home screen, the BrightcovePlayerView.java does not get destroyed.
This fix unregisters the listeners when the video view is detached (no longer on the screen), so BrightcovePlayerView.java. won't be listening to any events and avoids this problem.